Measure lexical variety

add_lexical_variety() augments a tidy text table with columns describing the lexical variety of the corpus. Among other things, checks for uniqueness and size of vocabulary, with additional ratios reporting these measurements in relation to document size.

Usage

add_lexical_variety(df, by = doc_id, feature = word)

Arguments

df

A tidy data frame, potentially containing columns called "doc_id" and "word"

by

A grouping column

feature

A column of words containing one word per row

Value

A data frame with 7 added columns , the first two logical and the rest numeric:

new_word (logical) Indicates whether this is the first instance of a given word

hapax (logical) Indicates whether this word is the only incident of a given word, or hapax legomenon

vocabulary (integer) Running count of words used

ttr (double) Type-token ratio, derived from the running count of words divided by the total number of words used

htr (double) Hapax-token ratio, derived from the running count of hapax legomena divided by the total number of words used

progress_words (integer) Running count of total words used so far in a document

progress_percent (double) Words used so far as a percentage of the total number of words used in a document

Examples

austen <- "austen.rds" |>
  system.file(package = "tmtyro") |>
  readRDS()

austen |>
   add_lexical_variety() |>
   head()
#> # A tibble: 6 × 12
#>   doc_id     par_num word   pos   lemma  new_word hapax vocabulary   ttr   htr
#>   <chr>        <dbl> <chr>  <chr> <chr>  <lgl>    <lgl>      <int> <dbl> <dbl>
#> 1 lady_susan       1 i      PRP   i      TRUE     FALSE          1     1     0
#> 2 lady_susan       2 lady   NNP   lady   TRUE     FALSE          2     1     0
#> 3 lady_susan       2 susan  NNP   susan  TRUE     FALSE          3     1     0
#> 4 lady_susan       2 vernon NNP   vernon TRUE     FALSE          4     1     0
#> 5 lady_susan       2 to     TO    to     TRUE     FALSE          5     1     0
#> 6 lady_susan       2 mr     NNP   mr     TRUE     FALSE          6     1     0
#> # ℹ 2 more variables: progress_words <int>, progress_percent <dbl>