Build and load a corpus from Project Gutenberg

get_gutenberg_corpus() improves upon the functionality of gutenbergr::gutenberg_download() in three key ways.

First, its default setting is to retrieve the ".htm" version of texts and parse the headers. This change allows for consideration of texts by sections and chapters. Parsing is handled by parse_html(), and move_header_to_text() is available to correct any parsing errors.

Second, when using the ".txt" format, get_gutenberg_corpus() doesn't rely on the presence of ".zip" files. This change improves coverage of files dramatically.

Finally, it caches a copy of files locally to avoid the need for repeated downloads. This change helps with code portability, offline access, and server bandwidth.

All changes are made with consideration for server bandwidth, so a two-second delay is introduced between each download attempt. This will slow down the initial acquisition of corpora, but offline caching speeds things up considerably in subsequent use.

Usage

get_gutenberg_corpus(
  gutenberg_id,
  dir = "gutenberg",
  mirror = NULL,
  meta_fields = c("title", "author"),
  type = c("htm", "txt"),
  html_title = FALSE,
  ...
)

Arguments

gutenberg_id

A vector of ID numbers from Project Gutenberg or a data frame containing a gutenberg_id column, such as from the results of a call to gutenbergr::gutenberg_works().

dir

The directory for storing downloaded .txt files. Default value is "gutenberg".

mirror

Optionally a mirror URL to retrieve the books from. By default uses the mirror from gutenbergr::gutenberg_get_mirror().

meta_fields

Additional fields to add from gutenbergr::gutenberg_metadata describing each book. By default, title and author are added.

type

Indicate whether the corpus should be created from the plain text "txt" versions of books (previously the only type) or from enhanced "htm" versions, which attempt to find markers for chapters and sections.

html_title

Whether to use the h1 header from an HTML file to determine a document's title. By default, uses gutenbergr::gutenberg_metadata.

...

Additional parameters passed along to gutenbergr::gutenberg_strip().

Value

A data frame with one row for each line of the texts in the corpus.

Examples

library(gutenbergr)

dalloway <- gutenberg_works(author == "Woolf, Virginia",
                            title == "Mrs Dalloway in Bond Street") |>
  get_gutenberg_corpus()