Load a folder of texts

load_texts() loads a corpus from a folder of texts and prepare it for further study using tidytext principles. By default, load_texts() will add paragraph numbers (suitable for prose), and unnest at the word level, but options exist to change these defaults for poetry, to avoid unnesting, and even to remove words that seem like proper nouns or to apply techniques of natural language processing for lemmatizing words or tagging their parts of speech.

Usage

load_texts(
  folder = "data",
  name = ".txt",
  word = TRUE,
  lemma = FALSE,
  lemma_replace = FALSE,
  to_lower = TRUE,
  remove_names = FALSE,
  pos = FALSE,
  poetry = FALSE,
  paragraph = TRUE
)

Arguments

folder

A directory containing prose texts. Defaults to "data".

name

What naming pattern to search for in this folder. Defaults to ".txt".

word

Whether to split one word per line. Defaults to TRUE.

lemma

Whether to lemmatize the text. When word is TRUE, adds a new column called lemma. This step can add a lot of time, so it defaults to FALSE.

lemma_replace

When lemma and word are both TRUE, toggles whether to replace the word column with the lemmatized tokens. Defaults to FALSE

to_lower

When word is TRUE, toggles whether to convert all words to lowercase. Defaults to TRUE.

remove_names

When word is TRUE, toggles whether to remove words that only appear with the form of initial capitals. Defaults to FALSE.

pos

Whether to add a column for part-of-speech tag. This step can add a lot of time, so it defaults to FALSE.

poetry

Whether to detect and indicate stanza breaks and line breaks. Defaults to FALSE.

paragraph

Whether to detect paragraph breaks for prose. Defaults to TRUE.

Value

A data frame with two to five columns and one row for each word (optionally, one row for each paragraph or one row for each line) in the corpus.

Examples

if (FALSE) {
mysteries <-
  load_texts("mystery-novels")

dickinson <-
  load_texts("dickinson-poems",
             poetry = TRUE)

# `load_texts()` can also be used with
# a traditional tidytext workflow:
mysteries <-
  load_texts("mystery-novels",
             word = FALSE,
             to_lower = FALSE) |>
  tidytext::unnest_tokens(word, text)
}